refactor: move reset-password DTOs to controller layer
Request/response records defined inside PasswordResetService forced
the controller to import types from the service layer, violating the
project's layered architecture. Moving them to AuthController keeps
them alongside LoginRequest/LoginResponse, which follow the same
pattern.

Add max = 128 to the password @Size constraint to prevent resource
exhaustion: Argon2 is memory-hard, so an unbounded input allows a
caller to force excessive CPU/memory consumption per request.
