+++ src/test/java/com/wcc/platform/controller/platform/AuthControllerTest.java
@@ -23,9 +23,9 @@ import com.wcc.platform.domain.platform.member.MemberDto;
 import com.wcc.platform.domain.platform.type.RoleType;
 import com.wcc.platform.service.AuthService;
 import com.wcc.platform.service.MemberService;
+import com.wcc.platform.controller.platform.AuthController.ConfirmPasswordResetRequest;
+import com.wcc.platform.controller.platform.AuthController.ResetPasswordRequest;
 import com.wcc.platform.service.PasswordResetService;
-import com.wcc.platform.service.PasswordResetService.ConfirmPasswordResetRequest;
-import com.wcc.platform.service.PasswordResetService.ResetPasswordRequest;
 import java.time.OffsetDateTime;
 import java.util.List;
 import java.util.Optional;
