+++ src/test/java/com/wcc/platform/service/PasswordResetServiceTest.java
@@ -24,7 +24,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -44,7 +43,7 @@ class PasswordResetServiceTest {
   @Mock private PasswordEncoder passwordEncoder;
   @Mock private EmailService emailService;
 
-  @InjectMocks private PasswordResetService passwordResetService;
+  private PasswordResetService passwordResetService;
 
   @BeforeEach
   void setUp() {
